Protect comments PUT and DELETE routes
In order to update or delete comments users must
be signed in. To delete a specific resource a
userid must be provided as a parameter as well as
an id for the comment. The userid provided as a
parameter must also match the id stored in the
JWT. This ensures the user can only delete or
update posts that belong to themselves.

+++ controllers/commentsController.js
@@ -88,6 +88,10 @@ const addComment = asyncHandler(async (req, res) => {
 
 const deleteComment = asyncHandler(async (req, res) => {
   try {
+    if (!req.user) {
+      throw new Error(`Must be logged in as a user to delete comments`, 403);
+    }
+
     const commentId = Number(req.params.commentid);
 
     await database.deleteComment(commentId);
@@ -103,8 +107,17 @@ const deleteComment = asyncHandler(async (req, res) => {
 
 const updateComment = asyncHandler(async (req, res) => {
   try {
+    if (!req.user) {
+      throw new Error(`Must be logged in as a user to update comments`, 403);
+    }
+
     const commentId = Number(req.params.commentid);
     const content = req.body.content;
+
+    const comment = await database.getComment(commentId);
+    if (!comment) {
+      throw new Error(`Unable to find comment`, 404);
+    }
 
     await database.updateComment(commentId, content);
